isolating problematic test
