+++ src/axolotl/utils/models.py
@@ -843,7 +843,6 @@ class ModelLoader:
             if self.cfg.is_multimodal:
                 self.model_config.text_config = self.text_model_config
 
-            # self.model._attn_implementation_autoset = False
             self.model = self.AutoModelLoader.from_pretrained(
                 self.base_model,
                 config=self.model_config,
